Changed quadratic_equation variables to float

+++ quadratic_equation.rb
@@ -1,11 +1,11 @@
 print "Пожалуйста введите a: "
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 
 print "Пожалуйста введите b: "
-b = gets.chomp.to_i
+b = gets.chomp.to_f
 
 print "Пожалуйста введите c: "
-c = gets.chomp.to_i
+c = gets.chomp.to_f
  
 d = b ** 2 - (4 * a * c)
